perlmutter: add a checkpoint file
to workaround very slow job turnaround

+++ run_mtt.sh
@@ -15,10 +15,15 @@ if [ $# -eq 0 ] ; then
 else
   BRANCH=$1
 fi
+if [ -f ./running_$BRANCH ] ; then
+  exit
+fi
+touch ./running_$BRANCH
 SCRATCH_FILE=$BRANCH"_scratch"
 SCRATCH_DIR=/global/homes/h/hpp/mtt_perlmutter/$SCRATCH_FILE
 rm -f -r $SCRATCH_DIR
 export MTT_HOME=$PWD
 echo "============== Testing $BRANCH  ==============="
-pyclient/pymtt.py --verbose  get_ompi_$BRANCH.ini
+pyclient/pymtt.py  --verbose get_ompi_$BRANCH.ini
+rm ./running_$BRANCH
